+++ components/src/core/drawer/drawer-nav-item.tsx
@@ -53,11 +53,9 @@ const makeStyles = (
     flipIcon: ViewStyle;
 }> => {
     // Primary color manipulation
-    // @ts-ignore
     const fivePercentOpacityPrimary = color(theme.colors.primaryBase || theme.colors.primary)
         .fade(0.95)
         .string();
-    // @ts-ignore
     const twentyPercentOpacityPrimary = color(theme.colors.primaryBase || theme.colors.primary)
         .fade(0.8)
         .string();
@@ -108,7 +106,6 @@ export const DrawerNavItem: React.FC<DrawerNavItemProps> = (props) => {
     const previousActive = usePrevious(activeItem || '');
 
     // approximating primary[200] but we don't have access to it directly from the theme
-    // @ts-ignore
     const lightenedPrimary = color(theme.colors.primaryBase || theme.colors.primary)
         .lighten(0.83)
         .desaturate(0.39)
@@ -246,7 +243,6 @@ export const DrawerNavItem: React.FC<DrawerNavItemProps> = (props) => {
                             divider={showDivider ? 'full' : undefined}
                             statusColor={statusColor}
                             fontColor={active ? activeItemFontColor : itemFontColor}
-                            // @ts-ignore
                             IconClass={icon}
                             iconColor={active ? activeItemIconColor : itemIconColor}
                             rightComponent={