+++ components/src/core/empty-state/empty-state.test.tsx
@@ -4,7 +4,10 @@ import { Button } from 'react-native-paper';
 import { wrapIcon, EmptyState } from '..';
 import MatIcon from 'react-native-vector-icons/MaterialIcons';
 import MatCommunityIcon from 'react-native-vector-icons/MaterialCommunityIcons';
-const Line = wrapIcon({ IconClass: MatCommunityIcon, name: 'chart-line-variant' });
+const Line = wrapIcon({
+    IconClass: MatCommunityIcon,
+    name: 'chart-line-variant',
+});
 
 describe('EmptyState Tests ', () => {
     it('Icon Renders', () => {