+++ components/src/core/header/header.tsx
@@ -41,7 +41,6 @@ const headerStyles = (
     return StyleSheet.create({
         root: {
             width: '100%',
-            // @ts-ignore
             backgroundColor: props.backgroundColor || theme.colors.primaryBase || theme.colors.primary,
             shadowColor: 'rgba(0, 0, 0, 0.3)',
             shadowOffset: {