+++ components/src/core/list-item-tag/list-item-tag.tsx
@@ -30,7 +30,6 @@ const listItemTagStyles = (
 }> =>
     StyleSheet.create({
         root: {
-            // @ts-ignore
             backgroundColor: props.backgroundColor || theme.colors.primaryBase || theme.colors.primary,
             color:
                 props.fontColor ||
@@ -42,9 +41,7 @@ const listItemTagStyles = (
             paddingLeft: 4 * fontScale,
             paddingRight: 3 * fontScale, // to account for the 1px letter spacing on the last letter
             borderRadius: 2 * fontScale,
-            // @ts-ignore TODO: bold font is not available from default RNP theming yet
             fontFamily: theme.fonts.bold ? theme.fonts.bold.fontFamily : theme.fonts.medium.fontFamily,
-            // @ts-ignore TODO: bold font is not available from default RNP theming yet
             fontWeight: theme.fonts.bold ? theme.fonts.bold.fontWeight : theme.fonts.medium.fontWeight,
             overflow: 'hidden',
             lineHeight: 16,